Fixed bug in integration tests. Fails to delete if brach directory doesn't exists

+++ MabadoIntegrationTests/ConnectionResolverIntegrationTests.cs
@@ -68,7 +68,8 @@ namespace ConnectionResolverIntegrationTests
 
         private static void DeleteBranchFolder()
         {
-            Directory.Delete("Branch", true);
+            if (Directory.Exists("Branch"))
+                Directory.Delete("Branch", true);
         }
 
         private static void DeployBranchFolder()
